refactor: rename from tosshin_dienen_controller into dienen_controller
