+++ .github/workflows/build-and-test-nightly.yml
@@ -12,7 +12,7 @@ jobs:
       - name: Checkout this repository
         uses: actions/checkout@v2.3.4
         with:
-          path: tosshin_dienen_controller
+          path: dienen_controller
 
       - name: Add nightly Debian repository and rosdep sources list
         run: |
