+++ .gitignore
@@ -3,8 +3,8 @@
 !.git*
 
 build
-log
 install
+log
 
 debian
 obj-x86_64-linux-gnu
