+++ README.md
@@ -1,13 +1,13 @@
-# Tosshin Dienen Controller
+# Dienen Controller
 
-[![latest version](https://img.shields.io/github/v/release/threeal/tosshin_dienen_controller)](https://github.com/threeal/tosshin_dienen_controller/releases/)
-[![commits since latest version](https://img.shields.io/github/commits-since/threeal/tosshin_dienen_controller/latest)](https://github.com/threeal/tosshin_dienen_controller/commits/master)
-[![license](https://img.shields.io/github/license/threeal/tosshin_dienen_controller)](./LICENSE)
-[![stable test status](https://img.shields.io/github/workflow/status/threeal/tosshin_dienen_controller/Build%20and%20Test%20Stable?label=stable%20test)](https://github.com/threeal/tosshin_dienen_controller/actions)
-[![nightly test status](https://img.shields.io/github/workflow/status/threeal/tosshin_dienen_controller/Build%20and%20Test%20Nightly?label=nighly%20test)](https://github.com/threeal/tosshin_dienen_controller/actions)
+[![latest version](https://img.shields.io/github/v/release/threeal/dienen_controller)](https://github.com/threeal/dienen_controller/releases/)
+[![commits since latest version](https://img.shields.io/github/commits-since/threeal/dienen_controller/latest)](https://github.com/threeal/dienen_controller/commits/master)
+[![license](https://img.shields.io/github/license/threeal/dienen_controller)](./LICENSE)
+[![stable test status](https://img.shields.io/github/workflow/status/threeal/dienen_controller/Build%20and%20Test%20Stable?label=stable%20test)](https://github.com/threeal/dienen_controller/actions)
+[![nightly test status](https://img.shields.io/github/workflow/status/threeal/dienen_controller/Build%20and%20Test%20Nightly?label=nighly%20test)](https://github.com/threeal/dienen_controller/actions)
 
-This package provides a [Dienen](https://github.com/threeal/proposal-ta-simulasi-robot) assistive robot's controller client implementation of [Tosshin](https://github.com/threeal/tosshin) mobile robot navigation project.
+This package provides controller programs of the [Dienen](https://github.com/threeal/buku-ta-simulasi-robot) mobile assistive robot.
 
 ## License
 
-This project is maintained by [Alfi Maulana](https://github.com/threeal) and licensed under the [MIT License](./LICENSE).
+This project is licensed under the [MIT License](./LICENSE).
